Nozomi: reactivate ALS sensor and hack
du that sony DASH is still not ready. I got it working all
only compass not....and one problem with the light sensor
which is reporting wrong values back to the camera. Need
to look deeper into it then. But solution is not so far...I
will keep it in gerrit at the moment.

Till this is solved we use the old ALS version with
prebuild sensors

cherrs

Change-Id: I99e31f31d2195ff53207ef83139cd68e1bf1b966